feat(Browser): Adds browser support
The `captureStackTrace` method is very specific to the v8 engine. Most other browsers were failing due to it.
This commit makes it work everywhere at the cost of a degraded experience on other browsers: stack trace are way less useful. That said, as most people are using Chrome for dev, the module still have interest even for frontend dev.

Would be could to find a more universal way to print the usual YError stack trace but I think the only solution is to create a custom method (`printStack()` for example.